prevent stalling all broadcasts when single connection is blocked
Previously a GossipChannel had one GossipSender per broadcast source
peer. That sender would stall when any of the next-hop connections
were blocked, causing broadcasts from the same source peer to be
accumulate (via GossipData.Merge) locally even though it could be sent
down the remaining unblocked connections.

We can't just have a GossipSender per connection, because broadcast
messages carry the source peer and hence only messages from the same
source peer can be merged.

The obvious fix would be for GossipChannels to have a GossipSender per
combination of source peer and connection. But that is potentially
O(n_peers * n_connections) GossipSenders, which in a fully connected
mesh is O(n_peers^2).

So instead we do the following:

1. We abandon the concept of GossipData. The gossip interfaces now
deal in encoded gossip. This simplifies the interface and its usage
quite a lot and also makes it more amendable to possible future
exposure via, say, HTTP.

1. We get rid of the per broadcast source GossipSenders and instead
send broadcast messages via the per connection GossipSenders.

2. Instead of single GossipData cell, GossipSenders contain a buffer
of ProtocolMsgs. This buffer has a dynamic bound - the aggregate size
of the encoded gossip must not exceed the size of the last complete
gossip times a certain factor (currently set to 2). When the size is
exceeded we clear the buffer and when the GossipSender is next ready
to send it obtains a complete gossip and sends that. More on that
below.

3. GossipSenders shut down when the sending fails. This ensures that
we waste no cycles obtaining complete gossip data for dead
connections.

The buffering logic is best considered as a compression scheme. There
are three types of gossip: complete, differential ("send on everything
new I've learnt from receiving some gossip") and incremental
(GossipBroadcast). There is no point buffering more than the complete
gossip size of differential gossip - we might as well just send the
complete gossip. The situation is the same for incremental gossip
except that replacing a bunch of incremental gossip with complete
gossip makes propagation less efficient, and can delay convergence,
since incremental gossip propagates via optimal broadcast routing
rather than neighbour gossiping. So we do give up something here. But
it's better than either having unbounded buffers, or quadratic space
complexity, or simply throwing away gossip when we cannot send it
immediately.

Note that we are assuming here that GossipBroadcast and Gossip talk
about the same information. That is the case in all current uses, and
documented in the Gossiper API by reference to a "state".

There is one wart here: SurrogateGossiper. It has no idea what
complete gossip sizes are since it does not generate any complete
gossip. So instead just buffers everything. Which is exactly what it
did previously, since it didn't know how to merge GossipData
either. So no change, really.

With the per-broadcast-source GossipSenders gone, we can get rid of
GossipSender gc...

Instead of maintaining a Connection->GossipSender map in the
GossipChannels - which we then need to gc when connections get closed
- we keep gossip_channel_name->GossipSender maps in LocalConnections.

Access to the map is guarded by a separate lock, rather than re-using
the LocalConnection lock. This reduces lock contention and risk of
deadlock.

In the GossipSenders, we replace the Stop() method with a signalling
channel, i.e. senders stop when that channel gets closed. Handily, the
LocalConnection.finished channel fits that description. So when a
LocalConnection's run loop terminates, any existing and future
associated GossipSenders will end up terminating too.