don't drop captured packets from remote MAC
We see such packets when a MAC that we previously saw at a remote peer
has been moved to ours. Dropping these packets cuts off network
connectivity for the new owner of the MAC.

Instead we update the MAC cache and flush the flows, to remove the now
incorrect entries for that MAC. And carry on as normal.