Add --netns-fd flag to weaveutil
The flag is used to denote that weaveutil has to enter the given
netns (identified by fd) before executing a cmd.

Keep in mind, that this is still not the safest way to run some
Go code in custom netns, because some runtime sched threads
might reside in the root netns.