Set arp_accept when configuring ARP cache
Enabling "arp_accept" makes fail-overs faster. This is so, because
it prevents the kernel from dropping unsolicited ARP requests.

The disadvantage of the setting is that containers / weave bridge
become prone to the ARP spoofing attacks.

Keep in mind, that the documentation about "arp_accept" is a bit
misleading in Documentation/networking/ip-sysctl.txt: regardless
whether an entry with IP addr == unsolicited ARP request IP addr
exists in the ARP table, the kernel might drop the request if the
last update of the entry is within LOCKTIME boundaries.