Fix up tests for DNS changes
Now that `weave dns-args` doesn't parse any flags we need to
call it or not call it, which `start_container` will do for us.

Also add back the part of the test where we run weaver with and
without DNS, which had become lost in 8b061fd555d

Also change a couple of tests to use proxy to start containers,
again to avoid needing `dns-args` to parse flags.